backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.0
email-validator==2.2.0
bcrypt==4.1.2